+++ office-addin/src/App.tsx
@@ -7,6 +7,7 @@ import {
 } from "@erato/frontend/library";
 import { t } from "@lingui/core/macro";
 
+import { CalendarDevProbe } from "./components/CalendarDevProbe";
 import { AddinChatPage } from "./pages/AddinChatPage";
 import { OfficeProvider, useOffice } from "./providers/OfficeProvider";
 import { OfficeThemeProvider } from "./providers/OfficeThemeProvider";
@@ -123,6 +124,7 @@ export default function App() {
                     </OutlookWrapper>
                   </AuthGate>
                   <Toaster placement="bottom-center" />
+                  <CalendarDevProbe />
                 </OutlookAuthProvider>
               </OfficeThemeProvider>
             </ApiProvider>
